Added tests and checks

+++ server/src/main/java/com/techelevator/tenmo/dao/JdbcTransactionDao.java
@@ -45,19 +45,35 @@ public class JdbcTransactionDao implements TransactionDao {
         return transactionList;
     }
 
-    public Transaction getTransaction(int transactionId) {
+    public Transaction getTransaction(int transactionId, String username) {
         Transaction transaction = null;
+        int userId = 0;
+        int actualId = 0;
+        String sqlUserId = "SELECT tenmo_user.user_id FROM transactions JOIN account ON transactions.user_id = account.user_id JOIN tenmo_user ON account.user_id = tenmo_user.user_id WHERE username = ?;";
+        SqlRowSet result2 = jdbcTemplate.queryForRowSet(sqlUserId, username);
+        if (result2.next()) {
+            userId = result2.getInt("user_id");
+        }
         String sql = "SELECT transaction_id, user_id, account_id, amount, target_id, status FROM transactions WHERE transaction_id = ?;";
         try {
             SqlRowSet result = jdbcTemplate.queryForRowSet(sql, transactionId);
-            if (result.next()) {
-                transaction = mapRowToTransaction(result);
+
+                if (result.next()) {
+                    actualId = result.getInt("user_id");
+                    System.out.println(actualId);
+                    System.out.println(userId);
+                   if (actualId == userId) {
+                       transaction = mapRowToTransaction(result);
+                   }
+                } else {
+                    throw new TenmoException("Hey! This isn't your transaction ID >:(!");
             }
-        } catch (CannotGetJdbcConnectionException e) {
-            throw new DaoException("Unable to connect to server or database", e);
-        } catch (BadSqlGrammarException e) {
-            throw new DaoException("SQL syntax error", e);
-        }
+
+            } catch(CannotGetJdbcConnectionException e){
+                throw new DaoException("Unable to connect to server or database", e);
+            } catch(BadSqlGrammarException e){
+                throw new DaoException("SQL syntax error", e);
+            }
         return transaction;
     }
 
@@ -90,13 +106,23 @@ public class JdbcTransactionDao implements TransactionDao {
 
 
 
-    public Transaction createTransaction(Transaction transaction){
+    public Transaction createTransaction(Transaction transaction, String username){
         Transaction newTransaction = null;
-        String sql = "INSERT INTO transactions (user_id, account_id, amount, target_id, status) " +
-                "VALUES (?, ?, ?, ?, ?) RETURNING transaction_id;";
-        int transactionId = jdbcTemplate.queryForObject(sql, int.class, transaction.getUser_id(), transaction.getAccount_id(), transaction.getAmount(),
-                transaction.getTarget_id(), transaction.getStatus());
-        newTransaction = getTransaction(transactionId);
+        int userId = 0;
+        String sqlUserId = "SELECT user_id FROM tenmo_user WHERE username = ?;";
+        SqlRowSet result2 = jdbcTemplate.queryForRowSet(sqlUserId, username);
+        if (result2.next()) {
+            userId = result2.getInt("user_id");
+        }
+        if (transaction.getUser_id() == userId) {
+            String sql = "INSERT INTO transactions (user_id, account_id, amount, target_id, status) " +
+                    "VALUES (?, ?, ?, ?, ?) RETURNING transaction_id;";
+            int transactionId = jdbcTemplate.queryForObject(sql, int.class, transaction.getUser_id(), transaction.getAccount_id(), transaction.getAmount(),
+                    transaction.getTarget_id(), transaction.getStatus());
+            newTransaction = getTransaction(transactionId, username);
+        } else {
+            throw new TenmoException("You can only add transactions to your user accounts!" + userId + " did not match " + transaction.getUser_id());
+        }
         return newTransaction;
     }
 
@@ -122,20 +148,31 @@ public class JdbcTransactionDao implements TransactionDao {
 
     public void denyRequest(int transactionId, String name){
         int userId = 0;
+        int transactionTargetId = 0;
         String sql1 = "SELECT user_id FROM tenmo_user WHERE username = ?;";
+        String sqlTargetUser = "SELECT target_id FROM transactions WHERE transaction_id = ?;";
         SqlRowSet result = jdbcTemplate.queryForRowSet(sql1, name);
+        SqlRowSet result2 = jdbcTemplate.queryForRowSet(sqlTargetUser, transactionId);
+
         if(result.next()){
             userId = result.getInt("user_id");
         }
-        String sql =  "DELETE FROM transactions WHERE transaction_id = ? AND target_id = ? AND status = 'Pending';";
-        try {
-            jdbcTemplate.update(sql, transactionId, userId);
-
-        } catch(Exception e){
-            throw new TenmoException("You are not authorized to deny this request.");
+        if (result2.next()) {
+            transactionTargetId = result2.getInt("target_id");
         }
-        String sql2 = "DELETE FROM transactions WHERE transaction_id = ? - 1;";
-         jdbcTemplate.update(sql2, transactionId);
+        if (userId == transactionTargetId) {
+            String sql = "DELETE FROM transactions WHERE transaction_id = ? AND target_id = ? AND status = 'Pending';";
+            try {
+                jdbcTemplate.update(sql, transactionId, userId);
+
+            } catch (Exception e) {
+                throw new TenmoException("You are not authorized to deny this request.");
+            }
+            String sql2 = "DELETE FROM transactions WHERE transaction_id = ? - 1;";
+            jdbcTemplate.update(sql2, transactionId);
+        } else {
+            throw new TenmoException("You cannot deny this request only the target user can!");
+        }
     }
